feat: upgrade JS constraint validation
add a new, custom message in index.html for
validation errors via a span below the input
fields, and remove <br> tags.

modify style.css to help space the modal fields
and style the spans.

modify validation code in main.js to reveal spans,
create a red border when validation criteria are
not met for individual inputs on submit, also have
a secondary error span message if input exists for
the pages input field but it is an invalid input,
and clear all spans/field coloring when the
modal is closed and reopened.